cloud_functions/email-yield-curve-update/requirements.txt
# Function dependencies, for example:
# package>=version
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
db-dtypes==1.3.0
fonttools==4.53.1
google-api-core==2.19.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.26.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.20.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.66.1
idna==3.10
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==1.26.4
oauthlib==3.2.2
packaging==24.1
pandas==1.5.3
pandas-gbq==0.19.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.1
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydata-google-auth==1.8.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
threadpoolctl==3.5.0
urllib3==2.2.3
xlrd==2.0.1

# google-cloud-bigquery
# google-cloud-bigquery-storage
# google-cloud-secret-manager
# xlrd
# pyarrow
# numpy==1.26.4
# db-dtypes
# pytz
# matplotlib
# scikit-learn
# pandas==1.5.3
# pandas-gbq==0.19.1
